feat: move from "com" to "ru"
