--- library/src/main/java/com/mintrocket/lib/mintpermissions/MintPermissionsManager.kt
@@ -1,7 +1,0 @@
-package com.mintrocket.lib.mintpermissions
-
-import androidx.activity.ComponentActivity
-
-interface MintPermissionsManager {
-    fun init(activity: ComponentActivity)
-}
