--- library/src/main/java/com/mintrocket/lib/mintpermissions/internal/models/RequestResult.kt
@@ -1,8 +1,0 @@
-package com.mintrocket.lib.mintpermissions.internal.models
-
-import com.mintrocket.lib.mintpermissions.models.MintPermissionMultipleResult
-
-internal data class RequestResult(
-    val request: Request,
-    val result: MintPermissionMultipleResult
-)
