--- library/src/main/java/com/mintrocket/lib/mintpermissions/models/MintPermission.kt
@@ -1,3 +1,0 @@
-package com.mintrocket.lib.mintpermissions.models
-
-typealias MintPermission = String
