--- library/src/main/java/com/mintrocket/lib/mintpermissions/models/MintPermissionAction.kt
@@ -1,16 +1,0 @@
-package com.mintrocket.lib.mintpermissions.models
-
-sealed class MintPermissionAction(open val permission: MintPermission) {
-
-    data class Granted(
-        override val permission: MintPermission
-    ) : MintPermissionAction(permission)
-
-    data class NeedsRationale(
-        override val permission: MintPermission
-    ) : MintPermissionAction(permission)
-
-    data class DeniedPermanently(
-        override val permission: MintPermission
-    ) : MintPermissionAction(permission)
-}
