--- library/src/main/java/com/mintrocket/lib/mintpermissions/models/MintPermissionMultipleResult.kt
@@ -1,5 +1,0 @@
-package com.mintrocket.lib.mintpermissions.models
-
-data class MintPermissionMultipleResult(
-    val results: List<MintPermissionResult>
-)
