--- library/src/main/java/com/mintrocket/lib/mintpermissions/models/MintPermissionResult.kt
@@ -1,6 +1,0 @@
-package com.mintrocket.lib.mintpermissions.models
-
-data class MintPermissionResult(
-    val status: MintPermissionStatus,
-    val action: MintPermissionAction?
-)
