+++ library/src/main/java/ru/mintrocket/lib/mintpermissions/MintPermissionsManager.kt
@@ -1,0 +1,7 @@
+package ru.mintrocket.lib.mintpermissions
+
+import androidx.activity.ComponentActivity
+
+interface MintPermissionsManager {
+    fun init(activity: ComponentActivity)
+}
