+++ library/src/main/java/ru/mintrocket/lib/mintpermissions/models/MintPermission.kt
@@ -1,0 +1,3 @@
+package ru.mintrocket.lib.mintpermissions.models
+
+typealias MintPermission = String
