+++ library/src/main/java/ru/mintrocket/lib/mintpermissions/models/MintPermissionAction.kt
@@ -1,0 +1,16 @@
+package ru.mintrocket.lib.mintpermissions.models
+
+sealed class MintPermissionAction(open val permission: MintPermission) {
+
+    data class Granted(
+        override val permission: MintPermission
+    ) : MintPermissionAction(permission)
+
+    data class NeedsRationale(
+        override val permission: MintPermission
+    ) : MintPermissionAction(permission)
+
+    data class DeniedPermanently(
+        override val permission: MintPermission
+    ) : MintPermissionAction(permission)
+}
