+++ library/src/main/java/ru/mintrocket/lib/mintpermissions/models/MintPermissionMultipleResult.kt
@@ -1,0 +1,5 @@
+package ru.mintrocket.lib.mintpermissions.models
+
+data class MintPermissionMultipleResult(
+    val results: List<MintPermissionResult>
+)
