+++ library/src/main/java/ru/mintrocket/lib/mintpermissions/models/MintPermissionResult.kt
@@ -1,0 +1,6 @@
+package ru.mintrocket.lib.mintpermissions.models
+
+data class MintPermissionResult(
+    val status: MintPermissionStatus,
+    val action: MintPermissionAction?
+)
